Implement verify function & remove `hasCapability`.
