+++ tests/attenuation.test.ts
@@ -1,13 +1,9 @@
-import * as capability from "../src/capability"
 import * as token from "../src/token"
-
-import { alice, bob, mallory } from "./fixtures"
 import { emailCapabilities, emailCapability } from "./capability/email"
 
-import { equalCanDelegate, hasCapability } from "../src/attenuation"
-import { REDELEGATE } from "../src/capability/ability"
-import { SUPERUSER } from "../src/capability/super-user"
+import { alice, bob, mallory } from "./fixtures"
 import { all } from "../src/util"
+
 
 describe("attenuation.emailCapabilities", () => {
 
@@ -154,331 +150,6 @@ describe("attenuation.emailCapabilities", () => {
         capability: aliceEmail
       }
     ])
-  })
-
-})
-
-
-
-describe("hasCapability", () => {
-
-  async function aliceEmailDelegationExample() {
-    // alice -> bob, bob -> mallory
-    // alice delegates access to sending email as her to bob
-    // and bob delegates it further to mallory
-    const leafUcan = await token.build({
-      issuer: alice,
-      audience: bob.did(),
-      capabilities: [ emailCapability("alice@email.com") ]
-    })
-
-    const ucan = await token.build({
-      issuer: bob,
-      audience: mallory.did(),
-      capabilities: [ emailCapability("alice@email.com") ],
-      proofs: [ token.encode(leafUcan) ]
-    })
-
-    return ucan
-  }
-
-  function nowInSeconds() {
-    return Math.floor(Date.now() / 1000)
-  }
-
-  function aliceCapInfo() {
-    const capabilityWithInfo = {
-      capability: emailCapability("alice@email.com"),
-      // we need to provide some information about who we think originally
-      // created/has the capability
-      // and for which interval in time we want to check for the capability.
-      info: {
-        originator: alice.did(),
-        expiresAt: nowInSeconds(),
-      }
-    }
-
-    return capabilityWithInfo
-  }
-
-  it("gets a capability", async () => {
-    const ucan = await aliceEmailDelegationExample()
-    const cap = await hasCapability(equalCanDelegate, aliceCapInfo(), ucan)
-
-    expect(cap).toBeTruthy()
-
-    if (!cap) return
-
-    expect(cap.info.originator).toEqual(alice.did())
-    expect(cap.capability.with.hierPart).toEqual("alice@email.com")
-  })
-
-  it("rejects an invalid escalation", async () => {
-    const ucan = await aliceEmailDelegationExample()
-
-    // unix timestamp in seconds
-    const nowInSeconds = Math.floor(Date.now() / 1000)
-
-    const capabilityWithInfo = {
-      capability: {
-        ...emailCapability("alice@email.com"),
-        can: SUPERUSER
-      },
-      // we need to provide some information about who we think originally
-      // created/has the capability
-      // and for which interval in time we want to check for the capability.
-      info: {
-        originator: alice.did(),
-        expiresAt: nowInSeconds,
-      },
-    }
-
-    const cap = await hasCapability(equalCanDelegate, capabilityWithInfo, ucan)
-
-    expect(cap).toEqual(false)
-  })
-
-  it("rejects for an invalid originator", async () => {
-    const ucan = await aliceEmailDelegationExample()
-    // unix timestamp in seconds
-    const nowInSeconds = Math.floor(Date.now() / 1000)
-
-    const capabilityWithInfo = {
-      capability: emailCapability("alice@email.com"),
-      // we need to provide some information about who we think originally
-      // created/has the capability
-      // and for which interval in time we want to check for the capability.
-      info: {
-        // an invalid originator
-        originator: bob.did(),
-        notBefore: nowInSeconds,
-        expiresAt: nowInSeconds
-      }
-    }
-
-    const cap = await hasCapability(equalCanDelegate, capabilityWithInfo, ucan)
-
-    expect(cap).toEqual(false)
-  })
-
-  it("rejects for an expired capability", async () => {
-    const ucan = await aliceEmailDelegationExample()
-    // unix timestamp in seconds. Will be after
-    const nowInSeconds = Math.floor(Date.now() / 1000)
-
-    const capabilityWithInfo = {
-      capability: emailCapability("alice@email.com"),
-      // we need to provide some information about who we think originally
-      // created/has the capability
-      // and for which interval in time we want to check for the capability.
-      info: {
-        // an invalid originator
-        originator: bob.did(),
-        notBefore: nowInSeconds,
-        expiresAt: nowInSeconds + 60 * 60 * 24 // expiry is older than it should be
-      }
-    }
-
-    const cap = await hasCapability(equalCanDelegate, capabilityWithInfo, ucan)
-
-    expect(cap).toEqual(false)
-  })
-
-  it("supports redelegation with a `prf:*` capability", async () => {
-    // alice -> bob
-    // alice delegates access to sending email as her to bob
-    // and bob delegates it further to mallory
-    const leafUcan = await token.build({
-      issuer: alice,
-      audience: bob.did(),
-      capabilities: [ emailCapability("alice@email.com") ]
-    })
-
-    const ucan = await token.build({
-      issuer: bob,
-      audience: mallory.did(),
-      capabilities: [ capability.prf(capability.superUser.SUPERUSER, REDELEGATE) ],
-      proofs: [ token.encode(leafUcan) ]
-    })
-
-    const cap = await hasCapability(equalCanDelegate, aliceCapInfo(), ucan)
-
-    expect(cap).toBeTruthy()
-
-    if (!cap) return
-
-    expect(cap.info.originator).toEqual(alice.did())
-    expect(cap.capability.with.hierPart).toEqual("alice@email.com")
-  })
-
-  it("supports redelegation with a `prf:1` capability", async () => {
-    // alice -> bob
-    // alice delegates access to sending email as her to bob
-    // and bob delegates it further to mallory
-    const leafUcanA = await token.build({
-      issuer: alice,
-      audience: bob.did(),
-      capabilities: [ emailCapability("ignore-me@email.com") ]
-    })
-
-    const leafUcanB = await token.build({
-      issuer: alice,
-      audience: bob.did(),
-      capabilities: [ emailCapability("alice@email.com") ]
-    })
-
-    const ucan = await token.build({
-      issuer: bob,
-      audience: mallory.did(),
-      capabilities: [ capability.prf(1, REDELEGATE) ],
-      proofs: [ token.encode(leafUcanA), token.encode(leafUcanB) ]
-    })
-
-    const cap = await hasCapability(equalCanDelegate, aliceCapInfo(), ucan)
-
-    expect(cap).toBeTruthy()
-
-    if (!cap) return
-
-    expect(cap.info.originator).toEqual(alice.did())
-    expect(cap.capability.with.hierPart).toEqual("alice@email.com")
-
-    // Test invalid `prf:0`
-    const faultyUcan = await token.build({
-      issuer: bob,
-      audience: mallory.did(),
-      capabilities: [ capability.prf(0, REDELEGATE) ],
-      proofs: [ token.encode(leafUcanA), token.encode(leafUcanB) ]
-    })
-
-    const capFaulty = await hasCapability(equalCanDelegate, aliceCapInfo(), faultyUcan)
-
-    expect(capFaulty).toBeFalsy()
-  })
-
-  it("rejects an improper `prf` redelegation", async () => {
-    // alice -> bob
-    // alice delegates access to sending email as her to bob
-    // and bob delegates it further to mallory
-    const leafUcan = await token.build({
-      issuer: alice,
-      audience: bob.did(),
-      capabilities: [ emailCapability("invalid@email.com") ]
-    })
-
-    const ucan = await token.build({
-      issuer: bob,
-      audience: mallory.did(),
-      capabilities: [ capability.prf(capability.superUser.SUPERUSER, REDELEGATE) ],
-      proofs: [ token.encode(leafUcan) ]
-    })
-
-    const cap = await hasCapability(equalCanDelegate, aliceCapInfo(), ucan)
-
-    expect(cap).toBeFalsy()
-  })
-
-  it("supports redelegation with a `my` capability", async () => {
-    // alice -> bob
-    // alice delegates access to sending email as her to bob
-    // and bob delegates it further to mallory
-    const leafUcan = await token.build({
-      issuer: alice,
-      audience: bob.did(),
-      capabilities: [ capability.my(capability.superUser.SUPERUSER) ]
-    })
-
-    const ucan = await token.build({
-      issuer: bob,
-      audience: mallory.did(),
-      capabilities: [ emailCapability("alice@email.com") ],
-      proofs: [ token.encode(leafUcan) ]
-    })
-
-    const cap = await hasCapability(equalCanDelegate, aliceCapInfo(), ucan)
-
-    expect(cap).toBeTruthy()
-
-    if (!cap) return
-
-    expect(cap.info.originator).toEqual(alice.did())
-  })
-
-  it("supports redelegation with a `my` & `as` capability", async () => {
-    // alice -> bob, bob -> mallory, mallory -> "someone"
-    // alice delegates access to sending email as her to bob
-    // and bob delegates it further to mallory
-    const leafUcan = await token.build({
-      issuer: alice,
-      audience: bob.did(),
-      capabilities: [ capability.my(capability.superUser.SUPERUSER) ]
-    })
-
-    const middleUcan = await token.build({
-      issuer: bob,
-      audience: mallory.did(),
-      capabilities: [ capability.as(alice.did(), SUPERUSER) ],
-      proofs: [ token.encode(leafUcan) ]
-    })
-
-    const ucan = await token.build({
-      issuer: mallory,
-      audience: "did:key:someone",
-      capabilities: [ emailCapability("alice@email.com") ],
-      proofs: [ token.encode(middleUcan) ]
-    })
-
-    const cap = await hasCapability(equalCanDelegate, aliceCapInfo(), ucan)
-
-    expect(cap).toBeTruthy()
-
-    if (!cap) return
-
-    expect(cap.info.originator).toEqual(alice.did())
-  })
-
-  it("rejects an improper `my` redelegation", async () => {
-    // alice -> bob
-    // alice delegates access to sending email as her to bob
-    // and bob delegates it further to mallory
-    const leafUcan = await token.build({
-      issuer: alice,
-      audience: bob.did(),
-      capabilities: []
-    })
-
-    const ucan = await token.build({
-      issuer: bob,
-      audience: mallory.did(),
-      capabilities: [ emailCapability("alice@email.com") ],
-      proofs: [ token.encode(leafUcan) ]
-    })
-
-    const cap = await hasCapability(equalCanDelegate, aliceCapInfo(), ucan)
-
-    expect(cap).toBeFalsy()
-  })
-
-  it("rejects an improper `as` redelegation - no `my`", async () => {
-    // alice -> bob
-    // alice delegates access to sending email as her to bob
-    // and bob delegates it further to mallory
-    const leafUcan = await token.build({
-      issuer: alice,
-      audience: bob.did(),
-      capabilities: [ capability.as(bob.did(), SUPERUSER) ]
-    })
-
-    const ucan = await token.build({
-      issuer: bob,
-      audience: mallory.did(),
-      capabilities: [ emailCapability("alice@email.com") ],
-      proofs: [ token.encode(leafUcan) ]
-    })
-
-    const cap = await hasCapability(equalCanDelegate, aliceCapInfo(), ucan)
-
-    expect(cap).toBeFalsy()
   })
 
 })
